+++ CHANGELOG.md
@@ -1,5 +1,47 @@
 # Changelog
 
+## v2.8.0 - 2026-05-18
+
+## [2.8.0](https://github.com/wenisch-tech/Kairos/compare/v2.7.6...v2.8.0) (2026-05-18)
+
+
+### Features
+
+* Introduces socks and HTTP Proxy support ([ff83c41](https://github.com/wenisch-tech/Kairos/commit/ff83c418ecb64b4e62fb85960795f552a566eb3a)), closes [#60](https://github.com/wenisch-tech/Kairos/issues/60) [#56](https://github.com/wenisch-tech/Kairos/issues/56)
+
+
+
+Docker image: ghcr.io/wenisch-tech/kairos:2.8.0
+
+
+## v2.7.6 - 2026-05-12
+
+### [2.7.6](https://github.com/wenisch-tech/Kairos/compare/v2.7.5...v2.7.6) (2026-05-12)
+
+
+### Build Systems
+
+* **deps:** bump org.springframework.ai:spring-ai-bom ([#62](https://github.com/wenisch-tech/Kairos/issues/62)) ([7417f85](https://github.com/wenisch-tech/Kairos/commit/7417f85ed651f6cc37d438f80f255d9a85da6d4c))
+
+
+
+Docker image: ghcr.io/wenisch-tech/kairos:2.7.6
+
+
+## v2.7.5 - 2026-05-05
+
+### [2.7.5](https://github.com/wenisch-tech/Kairos/compare/v2.7.4...v2.7.5) (2026-05-05)
+
+
+### Build Systems
+
+* **deps:** bump org.springframework.ai:spring-ai-bom ([#57](https://github.com/wenisch-tech/Kairos/issues/57)) ([07f9ca0](https://github.com/wenisch-tech/Kairos/commit/07f9ca099d2eacd11f3d33bdee11c3ce3472e59d))
+
+
+
+Docker image: ghcr.io/wenisch-tech/kairos:2.7.5
+
+
 ## v2.7.4 - 2026-05-05
 
 ### [2.7.4](https://github.com/wenisch-tech/Kairos/compare/v2.7.3...v2.7.4) (2026-05-05)